Fixed default parameter and derived parameters
I put the call to tweak_param_defaults after the inputs have been read
in from file because it is based on the read in values.

I updated which parameters are default and which are derived by
putting the derived ones in tweak_param_defaults and default ones in
set_param_defaults.

I changed all inputs to camel case for the purpose of uniformity.  For
the purpose of backword compatibility have it also set to read the old
way.