New Feature: Super Reptation
Super Reptation is a mc move which removes a bead from one end of a
polymer and adds one to the other.  The chemical identities are changed
inbetween so that the AB field is only affected at the ends.

Also combined moveTypes and nMoveTypes as they are always the same.

Also did some slight reorganization of move arguements.